Move parsing out into AST module

+++ src/sh/src/builtins.rs
@@ -1,8 +1,8 @@
 use shell;
 
-pub type Builtin = fn(&mut shell::Shell, &Vec<String>) -> Result<u8, ()>;
+pub type Builtin = fn(&mut shell::Shell, &Vec<String>) -> Result<i32, ()>;
 
-pub fn exit(shell: &mut shell::Shell, _argv: &Vec<String>) -> Result<u8, ()> {
+pub fn exit(shell: &mut shell::Shell, _argv: &Vec<String>) -> Result<i32, ()> {
     shell.exit(0);
     return Ok(0);
 }
